Move contents of /filters folder to /lib
