+++ client/app/components/items-list/components/ItemsTable.jsx
@@ -5,8 +5,7 @@ import classNames from "classnames";
 import Table from "antd/lib/table";
 import FavoritesControl from "@/components/FavoritesControl";
 import TimeAgo from "@/components/TimeAgo";
-import { durationHumanize } from "@/filters";
-import { formatDate, formatDateTime } from "@/filters";
+import { durationHumanize, formatDate, formatDateTime } from "@/lib/utils";
 
 // `this` refers to previous function in the chain (`Columns.***`).
 // Adds `sorter: true` field to column definition
